Refactor GameState
I wanted to refactor this some time ago. This new version is cleaner and don't pollute the window object. It's less verbose when using it and the "entities" array is better encapsulated.

All objects that use it have been changed to avoid multiple commits doing the same thing. Maybe there is a better way to do that than using a singleton? I have yet to discover it...